update geofenceby ID

+++ src/main/java/com/example/service/GeofenceService.java
@@ -6,12 +6,12 @@ import com.example.examplequerydslspringdatajpamaven.entity.Geofence;
 
 public interface GeofenceService {
 
-	public Set<Geofence> getAllGeofences(int id);
-	public Geofence getGeofenceById(int geofenceId);
-	public void deleteGeofence(int geofenceId);
-	public List<Geofence> checkDublicateGeofenceInAdd(int userId,String name);
-	public String addGeofence(Geofence geofence,int id);
-	public List<Geofence> checkDublicateGeofenceInEdit(int geofenceId,int userId,String name);
+	public Set<Geofence> getAllGeofences(Long id);
+	public Geofence getGeofenceById(Long geofenceId);
+	public void deleteGeofence(Long geofenceId);
+	public List<Geofence> checkDublicateGeofenceInAdd(Long userId,String name);
+	public String addGeofence(Geofence geofence,Long id);
+	public List<Geofence> checkDublicateGeofenceInEdit(Long geofenceId,Long userId,String name);
 	public void editGeofence(Geofence geofence);
